refactor(server): approval side-effects handler registry
Convert the ~280-line switch in applyApprovalSideEffect into a typed
APPROVAL_HANDLERS registry, modelled on MCP_ADAPTERS. Each approval type's
side effect moves verbatim into its own handler module under
services/approval-handlers/; the public applyApprovalSideEffect /
applyApprovalDeniedSideEffect signatures are unchanged, so resolve-approval
and existing tests are untouched. Adding a new approval side effect is now a
one-line registry addition plus a handler module.
